backend/src/requirements.txt
flask
mysql.connector
flask_cors
dotenv
flask_bcrypt
redis
